refactor: use subfolders for file management and persistence e2e tests
